+++ src/main/java/kitchenpos/menu/domain/MenuMapper.java
@@ -1,10 +1,6 @@
-package kitchenpos.menu.application;
+package kitchenpos.menu.domain;
 
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuProducts;
 import kitchenpos.menu.dto.MenuRequest;
-import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -12,12 +8,6 @@ import java.util.stream.Collectors;
 
 @Component
 public class MenuMapper {
-    private ProductRepository productRepository;
-
-    public MenuMapper(ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
     public Menu mapFrom(MenuRequest request) {
         return new Menu(
                 request.getName(),
