+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,6 +1,7 @@
 package kitchenpos.order.application;
 
 import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderMapper;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderValidator;
 import kitchenpos.order.dto.OrderRequest;
