+++ src/main/java/kitchenpos/order/domain/OrderMapper.java
@@ -1,9 +1,5 @@
-package kitchenpos.order.application;
+package kitchenpos.order.domain;
 
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderLineItems;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
 import org.springframework.stereotype.Component;
