+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -3,8 +3,7 @@ package kitchenpos.menu.application;
 import kitchenpos.common.Price;
 import kitchenpos.fixture.TestMenuRequestFactory;
 import kitchenpos.fixture.TestProductFactory;
-import kitchenpos.menu.application.MenuMapper;
-import kitchenpos.menu.application.MenuService;
+import kitchenpos.menu.domain.MenuMapper;
 import kitchenpos.menu.domain.*;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
